moved node filters from settings to main view
