+++ src/CSharp App/views/MainWindow.axaml
@@ -57,6 +57,7 @@
                     <ColumnDefinition Width="Auto" />
                     <ColumnDefinition Width="*" />
                 </Grid.ColumnDefinitions>
+                <Grid.RowDefinitions>Auto,Auto</Grid.RowDefinitions>
                 <TextBox
                     GotFocus="ResourceFilterTextBox_GotFocus"
                     Grid.Column="0"
@@ -67,7 +68,6 @@
                     Name="ResourceFilterTextBox"
                     ToolTip.Tip="{x:Static resources:ToolTips.ResourceFilter}"
                     Watermark="{Binding ResourcePathFilterCount, StringFormat='{}{0:N0} Resource Path Filters', FallbackValue='Error', UpdateSourceTrigger=PropertyChanged}">
-
                     <FlyoutBase.AttachedFlyout>
                         <Flyout Placement="BottomEdgeAlignedLeft" ShowMode="Transient">
                             <ScrollViewer MaxHeight="300">
@@ -160,6 +160,96 @@
                         </Flyout>
                     </FlyoutBase.AttachedFlyout>
                 </TextBox>
+                <Grid Grid.Row="1">
+                    <Grid.ColumnDefinitions>*, *</Grid.ColumnDefinitions>
+                    <Grid.RowDefinitions>Auto</Grid.RowDefinitions>
+                    <TextBlock
+                        Grid.Column="0"
+                        Grid.Row="0"
+                        HorizontalAlignment="Left"
+                        IsVisible="{Binding FilterSelectionVisibility}"
+                        Margin="5"
+                        Text="Node Filter"
+                        ToolTip.Tip="{x:Static resources:ToolTips.NodeFilter}"
+                        VerticalAlignment="Center" />
+                    <Button
+                        Background="#6b82b2"
+                        Grid.Column="1"
+                        Grid.Row="0"
+                        HorizontalAlignment="Stretch"
+                        HorizontalContentAlignment="Center"
+                        IsVisible="{Binding FilterSelectionVisibility}"
+                        Margin="5"
+                        VerticalContentAlignment="Center"
+                        x:Name="NodeTypeFilter">
+                        <Button.Content>
+                            <MultiBinding StringFormat="[ {0} / {1} ]">
+                                <Binding Path="CheckedCount" />
+                                <Binding Path="TotalCount" />
+                            </MultiBinding>
+                        </Button.Content>
+                        <Button.Styles>
+                            <Style Selector="Button:pointerover /template/ ContentPresenter">
+                                <Setter Property="Background" Value="#455472" />
+                            </Style>
+                        </Button.Styles>
+                        <Button.Flyout>
+                            <Flyout>
+                                <StackPanel>
+                                    <StackPanel
+                                        HorizontalAlignment="Center"
+                                        Margin="0,0,0,5"
+                                        Orientation="Horizontal">
+                                        <Button
+                                            Background="#6b82b2"
+                                            Click="SelectAllClick"
+                                            Content="{x:Static resources:Labels.SelectAll}"
+                                            HorizontalContentAlignment="Center"
+                                            Margin="5"
+                                            Width="100">
+                                            <Button.Styles>
+                                                <Style Selector="Button:pointerover /template/ ContentPresenter">
+                                                    <Setter Property="Background" Value="#455472" />
+                                                </Style>
+                                            </Button.Styles>
+                                        </Button>
+                                        <Button
+                                            Background="#6b82b2"
+                                            Click="DeselectAllClick"
+                                            Content="{x:Static resources:Labels.DeselectAll}"
+                                            HorizontalContentAlignment="Center"
+                                            Margin="5"
+                                            Width="100">
+                                            <Button.Styles>
+                                                <Style Selector="Button:pointerover /template/ ContentPresenter">
+                                                    <Setter Property="Background" Value="#455472" />
+                                                </Style>
+                                            </Button.Styles>
+                                        </Button>
+                                    </StackPanel>
+                                    <TextBox
+                                        HorizontalAlignment="Stretch"
+                                        Text="{Binding SearchQuery, UpdateSourceTrigger=PropertyChanged}"
+                                        Watermark="{x:Static resources:Watermarks.Search}" />
+                                    <ItemsControl ItemsSource="{Binding FilteredNodeTypeFilterItems}">
+                                        <ItemsControl.ItemTemplate>
+                                            <DataTemplate>
+                                                <StackPanel Orientation="Horizontal" Width="400">
+                                                    <CheckBox IsChecked="{Binding IsChecked}" x:Name="FilterItemCheckBox" />
+                                                    <Label
+                                                        Content="{Binding Label}"
+                                                        Margin="5,0,0,0"
+                                                        PointerPressed="Label_Click"
+                                                        VerticalAlignment="Center" />
+                                                </StackPanel>
+                                            </DataTemplate>
+                                        </ItemsControl.ItemTemplate>
+                                    </ItemsControl>
+                                </StackPanel>
+                            </Flyout>
+                        </Button.Flyout>
+                    </Button>
+                </Grid>
             </Grid>
             <Grid Grid.Row="4" Margin="5">
                 <Grid.ColumnDefinitions>
